Moved tests to a separate file
